[feat] Move firmware download and validation into a separate layer
This change moves the firmware download and validation code out of the
FmcAlias layer into a new layer.
