+++ common/src/boot_status.rs
@@ -14,7 +14,8 @@ Abstract:
 
 const IDEVID_BOOT_STATUS_BASE: u32 = 1;
 const LDEVID_BOOT_STATUS_BASE: u32 = 65;
-const FMCALIAS_BOOT_STATUS_BASE: u32 = 129;
+const FWPROCESSOR_BOOT_STATUS_BASE: u32 = 129;
+const FMCALIAS_BOOT_STATUS_BASE: u32 = 193;
 
 /// Statuses used by ROM to log dice derivation progress.
 #[repr(u32)]
@@ -40,20 +41,23 @@ pub enum RomBootStatus {
     LDevIdCertSigGenerationComplete = LDEVID_BOOT_STATUS_BASE + 4,
     LDevIdDerivationComplete = LDEVID_BOOT_STATUS_BASE + 5,
 
+    // Firmware Processor Statuses
+    FwProcessorDownloadImageComplete = FWPROCESSOR_BOOT_STATUS_BASE,
+    FwProcessorManifestLoadComplete = FWPROCESSOR_BOOT_STATUS_BASE + 1,
+    FwProcessorImageVerificationComplete = FWPROCESSOR_BOOT_STATUS_BASE + 2,
+    FwProcessorPopulateDataVaultComplete = FWPROCESSOR_BOOT_STATUS_BASE + 3,
+    FwProcessorExtendPcrComplete = FWPROCESSOR_BOOT_STATUS_BASE + 4,
+    FwProcessorLoadImageComplete = FWPROCESSOR_BOOT_STATUS_BASE + 5,
+    FwProcessorFirmwareDownloadTxComplete = FWPROCESSOR_BOOT_STATUS_BASE + 6,
+    FwProcessorComplete = FWPROCESSOR_BOOT_STATUS_BASE + 7,
+
     // FmcAlias Statuses
-    FmcAliasDownloadImageComplete = FMCALIAS_BOOT_STATUS_BASE,
-    FmcAliasManifestLoadComplete = FMCALIAS_BOOT_STATUS_BASE + 1,
-    FmcAliasImageVerificationComplete = FMCALIAS_BOOT_STATUS_BASE + 2,
-    FmcAliasPopulateDataVaultComplete = FMCALIAS_BOOT_STATUS_BASE + 3,
-    FmcAliasExtendPcrComplete = FMCALIAS_BOOT_STATUS_BASE + 4,
-    FmcAliasLoadImageComplete = FMCALIAS_BOOT_STATUS_BASE + 5,
-    FmcAliasFirmwareDownloadTxComplete = FMCALIAS_BOOT_STATUS_BASE + 6,
-    FmcAliasDeriveCdiComplete = FMCALIAS_BOOT_STATUS_BASE + 7,
-    FmcAliasKeyPairDerivationComplete = FMCALIAS_BOOT_STATUS_BASE + 8,
-    FmcAliasSubjIdSnGenerationComplete = FMCALIAS_BOOT_STATUS_BASE + 9,
-    FmcAliasSubjKeyIdGenerationComplete = FMCALIAS_BOOT_STATUS_BASE + 10,
-    FmcAliasCertSigGenerationComplete = FMCALIAS_BOOT_STATUS_BASE + 11,
-    FmcAliasDerivationComplete = FMCALIAS_BOOT_STATUS_BASE + 12,
+    FmcAliasDeriveCdiComplete = FMCALIAS_BOOT_STATUS_BASE,
+    FmcAliasKeyPairDerivationComplete = FMCALIAS_BOOT_STATUS_BASE + 1,
+    FmcAliasSubjIdSnGenerationComplete = FMCALIAS_BOOT_STATUS_BASE + 2,
+    FmcAliasSubjKeyIdGenerationComplete = FMCALIAS_BOOT_STATUS_BASE + 3,
+    FmcAliasCertSigGenerationComplete = FMCALIAS_BOOT_STATUS_BASE + 4,
+    FmcAliasDerivationComplete = FMCALIAS_BOOT_STATUS_BASE + 5,
 }
 
 impl From<RomBootStatus> for u32 {
