+++ rom/dev/src/flow/cold_reset/fmc_alias.rs
@@ -14,40 +14,33 @@ Abstract:
 --*/
 
 use super::crypto::{Crypto, Ecc384KeyPair};
-use super::dice::{DiceInput, DiceLayer, DiceOutput};
+use super::dice::{DiceInput, DiceOutput};
+use super::fw_processor::FwProcInfo;
 use super::x509::X509;
+use crate::cprintln;
 use crate::flow::cold_reset::{copy_tbs, TbsType};
 use crate::flow::cold_reset::{KEY_ID_CDI, KEY_ID_FMC_PRIV_KEY};
-use crate::fuse::log_fuse_data;
 use crate::print::HexBytes;
 use crate::rom_env::RomEnv;
-use crate::verifier::RomImageVerificationEnv;
-use crate::{cprint, cprintln, pcr};
 use caliptra_common::dice;
-use caliptra_common::fuse::FuseLogEntryId;
 use caliptra_common::RomBootStatus::*;
 use caliptra_drivers::{
-    okref, report_boot_status, Array4x12, CaliptraResult, ColdResetEntry4, ColdResetEntry48,
-    DataVault, Hmac384Data, Hmac384Key, KeyId, KeyReadArgs, Lifecycle, Mailbox, MailboxRecvTxn,
-    ResetReason, SocIfc, WarmResetEntry4, WarmResetEntry48,
+    okref, report_boot_status, Array4x12, CaliptraResult, Hmac384Data, Hmac384Key, KeyId,
+    KeyReadArgs, Lifecycle,
 };
 use caliptra_error::CaliptraError;
-use caliptra_image_types::{ImageManifest, IMAGE_BYTE_SIZE};
-use caliptra_image_verify::{ImageVerificationInfo, ImageVerificationLogInfo, ImageVerifier};
-use caliptra_x509::{FmcAliasCertTbs, FmcAliasCertTbsParams, NotAfter, NotBefore};
-use core::mem::ManuallyDrop;
-use zerocopy::{AsBytes, FromBytes};
-
-extern "C" {
-    static mut MAN1_ORG: u32;
-}
+use caliptra_x509::{FmcAliasCertTbs, FmcAliasCertTbsParams};
 
 #[derive(Default)]
 pub struct FmcAliasLayer {}
 
-impl DiceLayer for FmcAliasLayer {
+impl FmcAliasLayer {
     /// Perform derivations for the DICE layer
-    fn derive(env: &mut RomEnv, input: &DiceInput) -> CaliptraResult<DiceOutput> {
+    pub fn derive(
+        env: &mut RomEnv,
+        input: &DiceInput,
+        fw_proc_info: &FwProcInfo,
+    ) -> CaliptraResult<()> {
         cprintln!("[afmc] ++");
         cprintln!("[afmc] CDI.KEYID = {}", KEY_ID_CDI as u8);
         cprintln!("[afmc] SUBJECT.KEYID = {}", KEY_ID_FMC_PRIV_KEY as u8);
@@ -55,43 +48,6 @@ impl DiceLayer for FmcAliasLayer {
             "[afmc] AUTHORITY.KEYID = {}",
             input.auth_key_pair.priv_key as u8
         );
-
-        // Download the image
-        let mut txn = Self::download_image(&mut env.soc_ifc, &mut env.mbox)?;
-
-        // Load the manifest
-        let manifest = Self::load_manifest(&mut txn);
-        let manifest = okref(&manifest)?;
-
-        let mut venv = RomImageVerificationEnv {
-            sha256: &mut env.sha256,
-            sha384: &mut env.sha384,
-            sha384_acc: &mut env.sha384_acc,
-            soc_ifc: &mut env.soc_ifc,
-            ecc384: &mut env.ecc384,
-            data_vault: &mut env.data_vault,
-            pcr_bank: &mut env.pcr_bank,
-        };
-
-        // Verify the image
-        let info = Self::verify_image(&mut venv, manifest, txn.dlen());
-        let info = okref(&info)?;
-
-        Self::update_fuse_log(&info.log_info)?;
-
-        // populate data vault
-        Self::populate_data_vault(venv.data_vault, info);
-
-        // Extend PCR0
-        pcr::extend_pcr0(&mut venv, info)?;
-        report_boot_status(FmcAliasExtendPcrComplete.into());
-
-        // Load the image
-        Self::load_image(manifest, &mut txn)?;
-
-        // Complete the mailbox transaction indicating success.
-        txn.complete(true)?;
-        report_boot_status(FmcAliasFirmwareDownloadTxComplete.into());
 
         // We use the value of PCR0 as the measurement for deriving the CDI.
         let measurement = env.pcr_bank.read_pcr(caliptra_drivers::PcrId::PcrId0);
@@ -120,271 +76,13 @@ impl DiceLayer for FmcAliasLayer {
             subj_key_id,
         };
 
-        // if there is a valid value in the manifest for the not_before and not_after
-        // we take it from there.
-
-        let mut nb = NotBefore::default();
-        let mut nf = NotAfter::default();
-        let null_time = [0u8; 15];
-
-        if manifest.header.vendor_data.vendor_not_after != null_time
-            && manifest.header.vendor_data.vendor_not_before != null_time
-        {
-            nf.not_after = manifest.header.vendor_data.vendor_not_after;
-            nb.not_before = manifest.header.vendor_data.vendor_not_before;
-        }
-
-        //The owner values takes preference
-        if manifest.header.owner_data.owner_not_after != null_time
-            && manifest.header.owner_data.owner_not_before != null_time
-        {
-            nf.not_after = manifest.header.owner_data.owner_not_after;
-            nb.not_before = manifest.header.owner_data.owner_not_before;
-        }
-
         // Generate Local Device ID Certificate
-        Self::generate_cert_sig(env, info, input, &output, &nb.not_before, &nf.not_after)?;
+        Self::generate_cert_sig(env, input, &output, fw_proc_info)?;
 
         report_boot_status(FmcAliasDerivationComplete.into());
         cprintln!("[afmc] --");
 
-        Ok(output)
-    }
-}
-
-impl FmcAliasLayer {
-    /// Download firmware mailbox command ID.
-    const MBOX_DOWNLOAD_FIRMWARE_CMD_ID: u32 = 0x46574C44;
-
-    /// Download the image
-    ///
-    /// # Arguments
-    ///
-    /// * `env` - ROM Environment
-    ///
-    /// # Returns
-    ///
-    /// Mailbox transaction handle. This transaction is ManuallyDrop because we
-    /// don't want the transaction to be completed with failure until after
-    /// report_error is called. This prevents a race condition where the SoC
-    /// reads FW_ERROR_NON_FATAL immediately after the mailbox transaction
-    /// fails, but before caliptra has set the FW_ERROR_NON_FATAL register.
-    fn download_image<'a>(
-        soc_ifc: &mut SocIfc,
-        mbox: &'a mut Mailbox,
-    ) -> CaliptraResult<ManuallyDrop<MailboxRecvTxn<'a>>> {
-        soc_ifc.flow_status_set_ready_for_firmware();
-
-        cprint!("[afmc] Waiting for Image ");
-        loop {
-            if let Some(txn) = mbox.peek_recv() {
-                if txn.cmd() != Self::MBOX_DOWNLOAD_FIRMWARE_CMD_ID {
-                    cprintln!("Invalid command 0x{:08x} received", txn.cmd());
-                    txn.start_txn().complete(false)?;
-                    continue;
-                }
-
-                // Re-borrow mailbox to work around https://github.com/rust-lang/rust/issues/54663
-                let txn = mbox
-                    .peek_recv()
-                    .ok_or(CaliptraError::FMC_ALIAS_MAILBOX_STATE_INCONSISTENT)?;
-
-                // This is a download-firmware command; don't drop this, as the
-                // transaction will be completed by either report_error() (on
-                // failure) or by a manual complete call upon success.
-                let txn = ManuallyDrop::new(txn.start_txn());
-                if txn.dlen() == 0 || txn.dlen() > IMAGE_BYTE_SIZE as u32 {
-                    cprintln!("Invalid Image of size {} bytes" txn.dlen());
-                    return Err(CaliptraError::FMC_ALIAS_INVALID_IMAGE_SIZE);
-                }
-
-                cprintln!("");
-                cprintln!("[afmc] Received Image of size {} bytes" txn.dlen());
-                report_boot_status(FmcAliasDownloadImageComplete.into());
-                return Ok(txn);
-            }
-        }
-    }
-
-    /// Load the manifest
-    ///
-    /// # Returns
-    ///
-    /// * `Manifest` - Caliptra Image Bundle Manifest
-    fn load_manifest(txn: &mut MailboxRecvTxn) -> CaliptraResult<ImageManifest> {
-        let slice = unsafe {
-            let ptr = &mut MAN1_ORG as *mut u32;
-            core::slice::from_raw_parts_mut(ptr, core::mem::size_of::<ImageManifest>() / 4)
-        };
-
-        txn.copy_request(slice)?;
-
-        if let Some(result) = ImageManifest::read_from(slice.as_bytes()) {
-            report_boot_status(FmcAliasManifestLoadComplete.into());
-            Ok(result)
-        } else {
-            Err(CaliptraError::FMC_ALIAS_MANIFEST_READ_FAILURE)
-        }
-    }
-
-    /// Verify the image
-    ///
-    /// # Arguments
-    ///
-    /// * `env` - ROM Environment
-    fn verify_image(
-        venv: &mut RomImageVerificationEnv,
-        manifest: &ImageManifest,
-        img_bundle_sz: u32,
-    ) -> CaliptraResult<ImageVerificationInfo> {
-        let mut verifier = ImageVerifier::new(venv);
-        let info = verifier.verify(manifest, img_bundle_sz, ResetReason::ColdReset)?;
-
-        cprintln!(
-            "[afmc] Image verified using Vendor ECC Key Index {}",
-            info.vendor_ecc_pub_key_idx,
-        );
-        report_boot_status(FmcAliasImageVerificationComplete.into());
-        Ok(info)
-    }
-
-    /// Update the fuse log
-    ///
-    /// # Arguments
-    /// * `log_info` - Image Verification Log Info
-    ///
-    /// # Returns
-    /// * CaliptraResult
-    fn update_fuse_log(log_info: &ImageVerificationLogInfo) -> CaliptraResult<()> {
-        // Log VendorPubKeyIndex
-        log_fuse_data(
-            FuseLogEntryId::VendorPubKeyIndex,
-            log_info.vendor_ecc_pub_key_idx.as_bytes(),
-        )?;
-
-        // Log VendorPubKeyRevocation
-        log_fuse_data(
-            FuseLogEntryId::VendorPubKeyRevocation,
-            log_info.fuse_vendor_pub_key_revocation.bits().as_bytes(),
-        )?;
-
-        // Log ManifestFmcSvn
-        log_fuse_data(
-            FuseLogEntryId::ManifestFmcSvn,
-            log_info.fmc_log_info.manifest_svn.as_bytes(),
-        )?;
-
-        // Log ManifestFmcMinSvn
-        log_fuse_data(
-            FuseLogEntryId::ManifestFmcMinSvn,
-            log_info.fmc_log_info.manifest_min_svn.as_bytes(),
-        )?;
-
-        // Log FuseFmcSvn
-        log_fuse_data(
-            FuseLogEntryId::FuseFmcSvn,
-            log_info.fmc_log_info.fuse_svn.as_bytes(),
-        )?;
-
-        // Log ManifestRtSvn
-        log_fuse_data(
-            FuseLogEntryId::ManifestRtSvn,
-            log_info.rt_log_info.manifest_svn.as_bytes(),
-        )?;
-
-        // Log ManifestRtMinSvn
-        log_fuse_data(
-            FuseLogEntryId::ManifestRtMinSvn,
-            log_info.rt_log_info.manifest_min_svn.as_bytes(),
-        )?;
-
-        // Log FuseRtSvn
-        log_fuse_data(
-            FuseLogEntryId::FuseRtSvn,
-            log_info.rt_log_info.fuse_svn.as_bytes(),
-        )?;
         Ok(())
-    }
-
-    /// Load the image to ICCM & DCCM
-    ///
-    /// # Arguments
-    ///
-    /// * `env`      - ROM Environment
-    /// * `manifest` - Manifest
-    /// * `txn`      - Mailbox Receive Transaction
-    fn load_image(manifest: &ImageManifest, txn: &mut MailboxRecvTxn) -> CaliptraResult<()> {
-        cprintln!(
-            "[afmc] Loading FMC at address 0x{:08x} len {}",
-            manifest.fmc.load_addr,
-            manifest.fmc.size
-        );
-
-        let fmc_dest = unsafe {
-            let addr = (manifest.fmc.load_addr) as *mut u32;
-            core::slice::from_raw_parts_mut(addr, manifest.fmc.size as usize / 4)
-        };
-
-        txn.copy_request(fmc_dest)?;
-
-        cprintln!(
-            "[afmc] Loading Runtime at address 0x{:08x} len {}",
-            manifest.runtime.load_addr,
-            manifest.runtime.size
-        );
-
-        let runtime_dest = unsafe {
-            let addr = (manifest.runtime.load_addr) as *mut u32;
-            core::slice::from_raw_parts_mut(addr, manifest.runtime.size as usize / 4)
-        };
-
-        txn.copy_request(runtime_dest)?;
-
-        report_boot_status(FmcAliasLoadImageComplete.into());
-        Ok(())
-    }
-
-    /// Populate data vault
-    ///
-    /// # Arguments
-    ///
-    /// * `env`  - ROM Environment
-    /// * `info` - Image Verification Info
-    fn populate_data_vault(data_vault: &mut DataVault, info: &ImageVerificationInfo) {
-        data_vault.write_cold_reset_entry48(ColdResetEntry48::FmcTci, &info.fmc.digest.into());
-
-        data_vault.write_cold_reset_entry4(ColdResetEntry4::FmcSvn, info.fmc.svn);
-
-        data_vault.write_cold_reset_entry4(ColdResetEntry4::FmcLoadAddr, info.fmc.load_addr);
-
-        data_vault.write_cold_reset_entry4(ColdResetEntry4::FmcEntryPoint, info.fmc.entry_point);
-
-        data_vault.write_cold_reset_entry48(
-            ColdResetEntry48::OwnerPubKeyHash,
-            &info.owner_pub_keys_digest.into(),
-        );
-
-        data_vault.write_cold_reset_entry4(
-            ColdResetEntry4::VendorPubKeyIndex,
-            info.vendor_ecc_pub_key_idx,
-        );
-
-        data_vault.write_warm_reset_entry48(WarmResetEntry48::RtTci, &info.runtime.digest.into());
-
-        data_vault.write_warm_reset_entry4(WarmResetEntry4::RtSvn, info.runtime.svn);
-
-        data_vault.write_warm_reset_entry4(WarmResetEntry4::RtLoadAddr, info.runtime.load_addr);
-
-        data_vault.write_warm_reset_entry4(WarmResetEntry4::RtEntryPoint, info.runtime.entry_point);
-
-        // TODO: Need a better way to get the Manifest address
-        let slice = unsafe {
-            let ptr = &MAN1_ORG as *const u32;
-            ptr as u32
-        };
-
-        data_vault.write_warm_reset_entry4(WarmResetEntry4::ManifestAddr, slice);
-        report_boot_status(FmcAliasPopulateDataVaultComplete.into());
     }
 
     /// Derive Composite Device Identity (CDI) from FMC measurements
@@ -432,11 +130,9 @@ impl FmcAliasLayer {
     /// * `output` - DICE Output
     fn generate_cert_sig(
         env: &mut RomEnv,
-        info: &ImageVerificationInfo,
         input: &DiceInput,
         output: &DiceOutput,
-        not_before: &[u8; FmcAliasCertTbsParams::NOT_BEFORE_LEN],
-        not_after: &[u8; FmcAliasCertTbsParams::NOT_AFTER_LEN],
+        fw_proc_info: &FwProcInfo,
     ) -> CaliptraResult<()> {
         let auth_priv_key = input.auth_key_pair.priv_key;
         let auth_pub_key = &input.auth_key_pair.pub_key;
@@ -445,7 +141,7 @@ impl FmcAliasLayer {
         let flags = Self::make_flags(env.soc_ifc.lifecycle(), env.soc_ifc.debug_locked());
 
         let svn = env.data_vault.fmc_svn() as u8;
-        let fuse_svn = info.fmc.effective_fuse_svn as u8;
+        let fuse_svn = fw_proc_info.fmc_effective_fuse_svn as u8;
 
         // Certificate `To Be Signed` Parameters
         let params = FmcAliasCertTbsParams {
@@ -461,8 +157,8 @@ impl FmcAliasLayer {
             tcb_info_flags: &flags,
             tcb_info_fmc_svn: &svn.to_be_bytes(),
             tcb_info_fmc_svn_fuses: &fuse_svn.to_be_bytes(),
-            not_before,
-            not_after,
+            not_before: &fw_proc_info.fmc_cert_valid_not_before.value,
+            not_after: &fw_proc_info.fmc_cert_valid_not_after.value,
         };
 
         // Generate the `To Be Signed` portion of the CSR
