+++ rom/dev/src/flow/cold_reset/mod.rs
@@ -15,6 +15,7 @@ Abstract:
 mod crypto;
 mod dice;
 mod fmc_alias;
+mod fw_processor;
 mod idev_id;
 mod ldev_id;
 mod x509;
@@ -22,6 +23,7 @@ mod x509;
 use crate::fht;
 use crate::flow::cold_reset::dice::*;
 use crate::flow::cold_reset::fmc_alias::FmcAliasLayer;
+use crate::flow::cold_reset::fw_processor::FirmwareProcessor;
 use crate::flow::cold_reset::idev_id::InitDevIdLayer;
 use crate::flow::cold_reset::ldev_id::LocalDevIdLayer;
 use crate::{cprintln, rom_env::RomEnv};
@@ -40,7 +42,6 @@ extern "C" {
     static mut FMCALIAS_TBS_ORG: u8;
     static mut LDEVID_TBS_SIZE: u8;
     static mut FMCALIAS_TBS_SIZE: u8;
-
 }
 
 pub enum TbsType {
@@ -60,15 +61,19 @@ impl ColdResetFlow {
     pub fn run(env: &mut RomEnv) -> CaliptraResult<FirmwareHandoffTable> {
         cprintln!("[cold-reset] ++");
 
-        // Compose the three dice layers into one function
-        let dice_fn = compose_layers(
-            InitDevIdLayer::derive,
-            compose_layers(LocalDevIdLayer::derive, FmcAliasLayer::derive),
-        );
+        // Execute IDEVID layer
+        let idevid_layer_output = InitDevIdLayer::derive(env)?;
+        let ldevid_layer_input = dice_input_from_output(&idevid_layer_output);
 
-        let input = DiceInput::default();
+        // Execute LDEVID layer
+        let ldevid_layer_output = LocalDevIdLayer::derive(env, &ldevid_layer_input)?;
+        let fmc_layer_input = dice_input_from_output(&ldevid_layer_output);
 
-        let _ = dice_fn(env, &input)?;
+        // Download and validate firmware.
+        let fw_proc_info = FirmwareProcessor::process(env)?;
+
+        // Execute FMCALIAS layer
+        FmcAliasLayer::derive(env, &fmc_layer_input, &fw_proc_info)?;
 
         cprintln!("[cold-reset] --");
 
@@ -114,4 +119,12 @@ pub fn copy_tbs(tbs: &[u8], tbs_type: TbsType, env: &mut RomEnv) -> CaliptraResu
 
     dst[..tbs.len()].copy_from_slice(tbs);
     Ok(())
+}
+
+fn dice_input_from_output(dice_output: &DiceOutput) -> DiceInput {
+    DiceInput {
+        auth_key_pair: &dice_output.subj_key_pair,
+        auth_sn: &dice_output.subj_sn,
+        auth_key_id: &dice_output.subj_key_id,
+    }
 }
