+++ x509/src/lib.rs
@@ -27,34 +27,32 @@ pub use idevid_csr::{InitDevIdCsrTbs, InitDevIdCsrTbsParams};
 pub use ldevid_cert::{LocalDevIdCertTbs, LocalDevIdCertTbsParams};
 pub use rt_alias_cert::{RtAliasCertTbs, RtAliasCertTbsParams};
 
+#[derive(Debug)]
 pub struct NotBefore {
-    pub not_before: [u8; 15],
+    pub value: [u8; 15],
 }
 
 impl Default for NotBefore {
     fn default() -> Self {
         let not_before = "20230101000000Z";
-        let mut nb: NotBefore = NotBefore {
-            not_before: [0u8; 15],
-        };
+        let mut nb: NotBefore = NotBefore { value: [0u8; 15] };
 
-        nb.not_before.copy_from_slice(not_before.as_bytes());
+        nb.value.copy_from_slice(not_before.as_bytes());
         nb
     }
 }
 
+#[derive(Debug)]
 pub struct NotAfter {
-    pub not_after: [u8; 15],
+    pub value: [u8; 15],
 }
 
 impl Default for NotAfter {
     fn default() -> Self {
         let not_after = "99991231235959Z";
-        let mut nb: NotAfter = NotAfter {
-            not_after: [0u8; 15],
-        };
+        let mut nf: NotAfter = NotAfter { value: [0u8; 15] };
 
-        nb.not_after.copy_from_slice(not_after.as_bytes());
-        nb
+        nf.value.copy_from_slice(not_after.as_bytes());
+        nf
     }
 }
